change following endpoints in app.py
PATCH /sets/<set_id>/cards | update_cards(set_id) | "Given a json object of card ids and set_id url param from backend adds the card ids to a learn set json object in the Progress field of the User SQL table. If the number of cards in Progress corresponds to the total number of cards belonging to that set, those card ids get removed from the Progress field in the User SQL table. Change: Already learned cards do not get appended to progress so as not to have duplicates"

Changes to testing mode:

Now only selects testing questions of cards that have not yet been learned (not yet in Progress of user)